+++ art/art_param.py
@@ -4,7 +4,7 @@ from .text_dic1 import *
 from .text_dic2 import *
 from .art_dic import *
 
-VERSION = "3.2"  # pragma: no cover
+VERSION = "3.3"  # pragma: no cover
 FONT_SMALL_THRESHOLD = 50  # pragma: no cover
 FONT_MEDIUM_THRESHOLD = 100  # pragma: no cover
 FONT_LARGE_THRESHOLD = 200  # pragma: no cover
@@ -13,6 +13,17 @@ TEXT_XLARGE_THRESHOLD = 3  # pragma: no cover
 TEXT_LARGE_THRESHOLD = 7  # pragma: no cover
 TEXT_MEDIUM_THRESHOLD = 10  # pragma: no cover
 
+ART_TYPE_ERROR = "The 'artname' type must be str."
+ART_NAME_ERROR = "Invalid art name."
+NUMBER_TYPE_ERROR = "The 'number' type must be int."
+TEXT_TYPE_ERROR = "The 'text' type must be str."
+FONT_TYPE_ERROR = "The 'font' type must be str."
+CHR_IGNORE_TYPE_ERROR = "The 'chr_ignore' type must be bool."
+FILE_TYPE_ERROR = "The 'filename' type must be str."
+PRINT_STATUS_TYPE_ERROR = "The 'print_status' type must be bool."
+ART_ENVIRONMENT_WARNING = "[Warning] This art is not printable in this environment."
+FONT_ENVIRONMENT_WARNING = "[Warning] This font is not printable in this environment."
+PACKAGE_LOAD_WARNING = "[Warning] There is a problem loading the package 'coverage'."
 
 SMALL_WIZARD_FONT = [
     "contessa",
@@ -136,7 +147,21 @@ RANDOM_FILTERED_FONTS = [
     "tsalagi",
     "gauntlet",
     "flyn_sh",
-    "moscow"]
+    "moscow",
+    "mirror",
+    "mirror_flip",
+    "flip",
+    "dwhistled"]
+
+RANDOM_FILTERED_ARTS = [
+    "message2",
+    "love you",
+    "text decoration",
+    "message1",
+    "musical"]
+
+
+TEST_FILTERED_FONTS = ["mirror", "mirror_flip", "flip"]
 
 DESCRIPTION = '''ASCII art is also known as "computer text art".
 It involves the smart placement of typed special characters or
@@ -465,7 +490,10 @@ FONT_MAP = {"block": [block_dic, True], "banner": [banner_dic, False],  # pragma
             "green_be": [green_be_dic, False],
             "hades": [hades_dic, False],
             "heavy_me": [heavy_me_dic, False],
-            "nfi1": [nfi1_dic, False]
+            "nfi1": [nfi1_dic, False],
+            "flip": [flip_dic, False],
+            "mirror": [mirror_dic, False],
+            "mirror_flip": [mirror_flip_dic, False]
 
             }